refactoring tests which are marked as slow
